this needs no array
If you really only want to sum up the first n numbers, you can do this
in a for loop without needing an additional array.

Be careful though:
You can't count from 0 to n-1 like before, the for loop needs to go
from 1 to n to give the correct result.

For the print( join() ), the number list is created on the fly via the
(1..n) list creator.