this needs no for loop
According to the Gaußsche Summenformel[1], the sum of the first n
integers is

  n     n²+n
  ∑ k = ----
 k=1      2

So by defining the formula f(x) = (x²+x)/2 the sum can be calculated
directly and without the need for a loop.

This is very fast for larger numbers as it is O(1) instead of O(n).

[1] see http://de.wikipedia.org/wiki/Gau%C3%9Fsche_Summenformel