code duplication
Both the sum calculation and the print() statements appear twice in the code.

In most cases, this is a bad thing, because in bigger projects, the
repetitions won't be directly next to each other.  When one part is
changed or fixed, the other might be missed and the code behaves
differently.

By writing a subroutine for the code, it can be called from different
locations with different parameters and is guaranteed to behave
exactly the same.